add email importance query crate
