+++ rust/cloud-storage/email/Cargo.toml
@@ -51,6 +51,7 @@ axum-extra = { workspace = true, optional = true }
 chrono = { workspace = true, features = ["serde"] }
 doppleganger = { workspace = true, features = ["chrono", "uuid"] }
 either = { workspace = true }
+email_importance = { path = "../email_importance" }
 email_utils = { path = "../email_utils" }
 filter_ast = { path = "../filter_ast" }
 frecency = { path = "../frecency" }
@@ -79,6 +80,7 @@ utoipa = { workspace = true, optional = true }
 uuid = { workspace = true }
 
 [dev-dependencies]
+email_importance = { path = "../email_importance", features = ["testing"] }
 frecency = { path = "../frecency", features = ["outbound", "postgres"] }
 macro_db_migrator = { path = "../macro_db_migrator" }
 serde_json = { workspace = true }
