+++ rust/cloud-storage/email_importance/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+edition = "2024"
+name = "email_importance"
+publish = false
+version = "0.1.0"
+
+[features]
+testing = []
+
+[dependencies]
+anyhow = { workspace = true }
+sqlx = { workspace = true }
+tokio = { workspace = true }
+tracing = { workspace = true }
+uuid = { workspace = true }
